Add @Autowired private MockMvc mockMvc
Spring can handle the incoming HTTP request without starting a server at all, to test only the layer below that. That way, almost of the full stack is used, and code will be called in exactly the same way as if it were processing a real HTTP request but without the cost of starting the server. MockMvc is used to do that.