Add testEmployerControllerReturnsEmployerPageView()
To perform this request Employer should be saved in jpaEmployerRepository.
To get employer’s id test uses the same repository. After employer is persisted it can be returned with id. Performing get all employers id is not needed. Then in this case a list of all employers – all - is returned.
This list has only 1 employer, test performs all.get(0) to get employer. 0 is first element in this list because it is an array list under the hood.
Rest of the test is a usual mockMvc test case.
MockMvc performs get “/employers/employer/{id}”, expects status 200 is ok, expects view “employer” with content “Employer”.