Add testVacancyControllerReturnsVacancyPageView()
To perform this request Vacancy should be saved in jpaVacancyRepository.
To get vacancy’s id test uses the same repository. After vacancy is persisted it can be returned with id.
To make that let's perform findAll(). Then in this case a list of all vacancies – all - is returned.
This list has only 1 vacancy, test performs all.get(0) to get vacancy. 0 is first element in this list because it is an
 array list under the hood.
Rest of the test is a usual mockMvc test case.
MockMvc performs get “/vacancies/vacancy/{id}”, expects status 200 is ok, expects view “vacancy” with content “Vacancy”.