Add testResumeControllerReturnsResumePageView()
To perform this request Resume should be saved in jpaResumeRepository.
To get resume’s id test uses the same repository. After resume is persisted it can be returned with id.
To make that let's perform findAll(). Then in this case a list of all resumes – all - is returned.
This list has only 1 resume, test performs all.get(0) to get resume. 0 is first element in this list because it is an
array list under the hood.
Rest of the test is a usual mockMvc test case.
MockMvc performs get “/resumes/resume/{id}”, expects status 200 is ok, expects view “resume” with content “Resume”.