Add testCoverLetterControllerReturnsCoverLetterPageView()
To perform this request CoverLetter should be saved in jpaCoverLetterRepository.
To get coverLetter’s id test uses the same repository. After coverLetter is persisted it can be returned with id.
To make that let's perform findAll(). Then in this case a list of all coverLetters – all - is returned.
This list has only 1 coverLetter, test performs all.get(0) to get coverLetter. 0 is first element in this list because it is an
array list under the hood.
Rest of the test is a usual mockMvc test case.
MockMvc performs get “/coverLetters/coverLetter/{id}”, expects status 200 is ok, expects view “cover_letter” with
content “Cover Letter”.